dela friendspanel.tsx
